Reworked interation test location and additional new code
