+++ i18n/sample/src/USSTree.i18n.json
@@ -1,5 +1,6 @@
 {
     "Favorites": "Favorites",
+    "renameUSSNode.error": "Unable to rename node: ",
     "enterPattern.log.debug.prompt": "Prompting the user to choose a member from the filtered list",
     "filterPrompt.log.debug.promptUSSPath": "Prompting the user for a USS path",
     "ussTree.error": "Error encountered in ",
