+++ i18n/sample/src/extension.i18n.json
@@ -54,6 +54,9 @@
     "showDSAttributes.log.error": "Error encountered when listing attributes! ",
     "showDSAttributes.error": "Unable to list attributes: ",
     "attributes.title": "Attributes",
+    "renameDataSet.log.debug": "Renaming data set ",
+    "renameDataSet.log.error": "Error encountered when renaming data set! ",
+    "renameDataSet.error": "Unable to rename data set: ",
     "hMigrate.requestSent1": "Migration of dataset: ",
     "hMigrate.requestSent2": " requested.",
     "renameDataSet.name": "Name of Data Set Member",
