+++ i18n/sample/src/uss/ussNodeActions.i18n.json
@@ -2,6 +2,5 @@
     "createUSSNode.name": "Name of file or directory",
     "createUSSNode.error.create": "Unable to create node: ",
     "ussNodeActions.error": "Error encountered in ",
-    "renameUSSNode.error": "Unable to rename node: ",
     "copyPath.infoMessage": "Copy Path is not yet supported in Theia."
 }
