+++ src/extension.ts
@@ -1115,60 +1115,6 @@ export async function showDSAttributes(parent: IZoweDatasetTreeNode, datasetProv
 
 }
 
-/**
- * Rename data sets
- *
- * @export
- * @param {IZoweDatasetTreeNode} node - The node
- * @param {DatasetTree} datasetProvider - the tree which contains the nodes
- */
-export async function renameDataSet(node: IZoweDatasetTreeNode, datasetProvider: IZoweTree<IZoweDatasetTreeNode>) {
-    let beforeDataSetName = node.label.trim();
-    let favPrefix = "";
-    let isFavourite;
-
-    if (node.contextValue.includes(FAV_SUFFIX)) {
-        isFavourite = true;
-        favPrefix = node.label.substring(0, node.label.indexOf(":") + 2);
-        beforeDataSetName = node.label.substring(node.label.indexOf(":") + 2);
-    }
-    const afterDataSetName = await vscode.window.showInputBox({value: beforeDataSetName});
-    const beforeFullPath = getDocumentFilePath(node.getLabel(), node);
-    const closedOpenedInstance = await closeOpenedTextFile(beforeFullPath);
-
-    log.debug(localize("renameDataSet.log.debug", "Renaming data set ") + afterDataSetName);
-    if (afterDataSetName) {
-        try {
-            await ZoweExplorerApiRegister.getMvsApi(node.getProfile()).renameDataSet(beforeDataSetName, afterDataSetName);
-            node.label = `${favPrefix}${afterDataSetName}`;
-
-            if (isFavourite) {
-                const profile = favPrefix.substring(1, favPrefix.indexOf("]"));
-                datasetProvider.renameNode(profile, beforeDataSetName, afterDataSetName);
-            } else {
-                const temp = node.label;
-                node.label = "[" + node.getSessionNode().label.trim() + "]: " + beforeDataSetName;
-                datasetProvider.renameFavorite(node, afterDataSetName);
-                node.label = temp;
-            }
-            datasetProvider.refreshElement(node);
-            datasetProvider.updateFavorites();
-
-            if (fs.existsSync(beforeFullPath)) {
-              fs.unlinkSync(beforeFullPath);
-            }
-
-            if (closedOpenedInstance) {
-                vscode.commands.executeCommand("zowe.ZoweNode.openPS", node);
-            }
-        } catch (err) {
-            log.error(localize("renameDataSet.log.error", "Error encountered when renaming data set! ") + JSON.stringify(err));
-            await utils.errorHandling(err, favPrefix, localize("renameDataSet.error", "Unable to rename data set: ") + err.message);
-            throw err;
-        }
-    }
-}
-
 function getProfileAndDataSetName(node: IZoweNodeType) {
     let profileName;
     let dataSetName;
@@ -1270,64 +1216,6 @@ export async function pasteDataSet(node: IZoweDatasetTreeNode, datasetProvider: 
             }
         } else {
             refreshPS(node);
-        }
-    }
-}
-
-/**
- * Rename data set members
- *
- * @export
- * @param {IZoweTreeNode} node - The node
- * @param {DatasetTree} datasetProvider - the tree which contains the nodes
- */
-export async function renameDataSetMember(node: IZoweTreeNode, datasetProvider: IZoweTree<IZoweDatasetTreeNode>) {
-    const beforeMemberName = node.label.trim();
-    let dataSetName;
-    let profileLabel;
-
-    if (node.getParent().contextValue.includes(FAV_SUFFIX)) {
-        profileLabel = node.getParent().getLabel().substring(0, node.getParent().getLabel().indexOf(":") + 2);
-        dataSetName = node.getParent().getLabel().substring(node.getParent().getLabel().indexOf(":") + 2);
-    } else {
-        dataSetName = node.getParent().getLabel();
-    }
-    const afterMemberName = await vscode.window.showInputBox({value: beforeMemberName});
-    const beforeFullPath = getDocumentFilePath(`${node.getParent().getLabel()}(${node.getLabel()})`, node);
-    const closedOpenedInstance = await closeOpenedTextFile(beforeFullPath);
-
-    log.debug(localize("renameDataSet.log.debug", "Renaming data set ") + afterMemberName);
-    if (afterMemberName) {
-        try {
-            await ZoweExplorerApiRegister.getMvsApi(node.getProfile()).renameDataSetMember(dataSetName, beforeMemberName, afterMemberName);
-            node.label = afterMemberName;
-        } catch (err) {
-            log.error(localize("renameDataSet.log.error", "Error encountered when renaming data set! ") + JSON.stringify(err));
-            await utils.errorHandling(err, profileLabel, localize("renameDataSet.error", "Unable to rename data set: ") + err.message);
-            throw err;
-        }
-        let otherParent;
-
-        if (node.getParent().contextValue.includes(FAV_SUFFIX)) {
-            otherParent = datasetProvider.findNonFavoritedNode(node.getParent());
-        } else {
-            otherParent = datasetProvider.findFavoritedNode(node.getParent());
-        }
-        if (otherParent) {
-            const otherMember = otherParent.children.find((child) => child.label === beforeMemberName);
-            if (otherMember) {
-                otherMember.label = afterMemberName;
-                datasetProvider.refreshElement(otherMember);
-            }
-        }
-        datasetProvider.refreshElement(node);
-
-        if (fs.existsSync(beforeFullPath)) {
-            fs.unlinkSync(beforeFullPath);
-        }
-
-        if (closedOpenedInstance) {
-            vscode.commands.executeCommand("zowe.ZoweNode.openPS", node);
         }
     }
 }
